+++ packages/hippo_auth/lib/src/auth_bloc.dart
@@ -3,6 +3,7 @@ import 'package:hippo_auth/hippo_auth.dart';
 import 'package:hippo_auth/src/auth_login_controller.dart';
 import 'package:hippo_core/hippo_core.dart';
 import 'package:hippo_core_flutter/hippo_core_flutter.dart';
+import 'package:http/http.dart' as http;
 
 class HippoAuthBloc extends BlocBase {
   final HippoAuthApiController apiController;
@@ -17,12 +18,14 @@ class HippoAuthBloc extends BlocBase {
     required Uri baseUrl,
     required KeyValueStore sessionStore,
     String? sessionKey,
+    http.Client? httpClient,
   }) {
     return HippoAuthBloc(
       apiController: HippoAuthApiController(
         baseUrl: baseUrl,
         sessionStore: sessionStore,
         sessionKey: sessionKey,
+        httpClient: httpClient,
       ),
     );
   }